+++ src/features/musicRecommendation/ui/index.tsx
@@ -1,0 +1,1 @@
+export { MusicRecommendationForm } from './MusicRecommendationForm';
