--- src/shared/ui/Create/index.tsx
@@ -1,12 +1,0 @@
-import { CreateForm } from './CreateForm';
-import { Container, Layout } from './styled';
-
-export const CreatePage = () => {
-  return (
-    <Layout>
-      <Container>
-        <CreateForm />
-      </Container>
-    </Layout>
-  );
-};
